refactor/optimize for major speed increases
Originally wrote this as a toy project in an afternoon, someone recently
brought it to my attention again, and I could see the code wasn't quite
up to my standards. Killed some nonsense in the hot loop, and used
lookup tables for each sine frequency. Didn't do an offical performance
test, but the speedup is very significant.